pyproject.toml
[project]
name = "otovdo"
version = "0.1.0"
description = "otomatically create short vdos"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "moviepy>=2.1.2",
    "niquests>=3.14.0",
    "python-dotenv>=1.1.0",
    "srt-equalizer>=0.1.10",
]
